chore: rely on renovate for security checks
There is no way to silence vulncheck when there is a non-relevant
security error (https://github.com/golang/go/issues/61211).

This is problematic when fixing such an error would require upgrading
a large amount of dependencies, for instance in the case of

https://github.com/ClickHouse/ch-go/security/advisories/GHSA-m454-3xv7-qj85

which is only ever relevant for testing an not production in the
context of Forgejo.

Now that renovate is used for stable branches, it can be used as an
alternative. It will propose relevant security updates by default and
it will also be possible to decline them if they do not matter.