chore: replace `github.com/go-testfixtures/testfixtures`
- Replaces `github.com/go-testfixtures/testfixtures` with a homebrew
solution that is fully compatible.
- The reason to replace this library is that it pulls in a lot of other
libraries which is causing issues: (1) the test binary becomes bigger
than necessary which really shows in incremental build times (this patch
removes 27.6MiB of the integration test binary) (2) it pulls in
libraries (mainly database drivers) that are not used and are not easy
to upgrade in case of a security vulnerability, causing CI failures.