feat: Patch testing on bare repository
- `testPatch` is a function that is called to test a pull request and
determine the state of the pull request. Checking for merge conflicts,
check if the diff is empty and if the pull request modify any protected
files.
- The checking for merge conflict and if the diff is empty is using git
commands that relied on a working tree to correctly functions. Forgejo
store repositories in a bare format which do not contain a working tree.
This means that a temporary copy was created every time a pull request
had to be re-checked and for large repositories involve quite some I/O
interaction.
- This patch adjust those codepaths to instead use newer Git porchelain
commands that work without requiring a work tree and can thus be used
directly on the bare repository. The merge conflict is now done via
[`git-merge-tree(1)`](https://git-scm.com/docs/git-merge-tree/) and
checking if the diff is empty is done via
[`git-diff-tree(1)`](https://git-scm.com/docs/git-diff-tree).
- If the function is called to test a patch where the head and base
repository are not the same, then [Git alternate](https://git-scm.com/docs/gitglossary#Documentation/gitglossary.txt-aiddefalternateobjectdatabaseaalternateobjectdatabase)
is used to make the head commit available in the base repository, this
done on a per git command basis via the
`GIT_ALTERNATE_OBJECT_DIRECTORIES` environment.
- As far as I can understand the documentation and the existing code,
there's no edge case that the new code cannot handle. It also results in
a cleaner codepath, as the existing code did a lot of checking and
merging in a more traditional approach that required a lot of (parsing)
code, while the new code offloads this to git and has a trivial parser
of the output.
- Resolves forgejo/forgejo#7701
- Added exhaustive integration testing.