chore: Fix outdated usage of unittest.OverrideFixtures
The function unittest.OverrideFixtures was changed with 7e489eed25f65a022ffc6503d74930943bee3a89.
The PR with the util_misc_test was older and merged after the change to
the unittest.OverrideFixtures function. So the test was still using the
old function, which leads to a conflict / error when running tests.