fix: remove all instances of `outline: none`
focus outlines provide an important interaction affordance; one that is particularly useful for power users and accessibility. it’s often removed to “clean up” design while in fact it detracts from the design by make it less useful. this is the first step in correcting this and making the design and interface more usable.