fix: `overflow-wrap` strategy in `.markup` CSS class (#7945)
This has been introduced in the cherry-picked commit from https://github.com/go-gitea/gitea/pull/34072

To me, adding this seems to be a bug as it forces a hard stop to any overflow nested within a markup div.

Why should it be removed again?
Because otherwise (wide) markdown tables might look super odd because the column breaking/wrapping logic is strange (likely related to the "anywhere" strategy being used).

- [Example of current state with `overflow-wrap`](https://codefloe.com/devxy/helm-rdepot/src/branch/main/charts/rdepot)
- [Example with `overflow-wrap` being removed](https://codefloe.com/devxy/helm-rdepot/src/branch/main/charts/rdepot)

| Current  | Patched  |
|---------|---------|
| ![image](/attachments/d4a32804-858c-4ed7-838a-30b969bbbf12) | ![image](/attachments/a1261dc2-da25-4d10-b46f-03967fcf57bb) |

The initial change in Gitea was motivated to resolve issues of content overflow caused by the expandable file list tree view they introduced lately. However, this seem to have caused the whole README to overflow in their case, which doesn't apply to Forgejo.

When removing this in Forgejo, only large markdown content (i.e. tables) are allowed to overflow. The whole README div will not overflow (I tested this with an extreme markdown table case).
These tables are then horizontally scrollable, similar to how GitHub handles this.

If the expandable tree view is going to be added to Forgejo, this should then be handled entirely differently. This specific change is not a solution.

I've also checked other value settings to `overflow-wrap` but none had any or any desirable effect.
I did not check all other possible CSS-related options for overflow*, but this might as well be a different task per se.

For the reasons outlined above, I am quite certain this change should be reverted (again).

Reviewed-on: https://codeberg.org/forgejo/forgejo/pulls/7945
Reviewed-by: Gusted <gusted@noreply.codeberg.org>
Reviewed-by: Otto <otto@codeberg.org>
Co-authored-by: pat-s <patrick.schratz@gmail.com>
Co-committed-by: pat-s <patrick.schratz@gmail.com>