feat: replace go-rpmutils library with our own
- While doing cross-compiling of Forgejo (with CGO enabled) it was
noticed that besides compiling sqlite3, github.com/DataDog/zstd (a CGO
wrapper around the reference zstd library) took a long time to compile.
Upon investigating why this library was included in the first place I
concluded that this library is not even used and compiling this package
is a waste of time and CPU cycles.
- github.com/sassoftware/go-rpmutils is the library that uses the
CGO zstd library, and would use it the pure Go
variant (github.com/klauspost/compress/zstd) if CGO is enabled (which is
the default). It uses zstd to uncompress the payload of a RPM packages.
This is a operation that Forgejo does not use in the slightest, hence
being unused code.
- It is not possible to force compiling the pure Go variant if CGO is
enabled. Therefore forking and removing this code is the only option to
avoid compiling the zstd C library. The changes made to the fork can be
seen here: https://code.forgejo.org/forgejo/go-rpmutils/compare/2660c86d578cd2a5366fe5bda488e37f9ab1b11a...v1.0.0
- Via [actiongraph](https://github.com/icio/actiongraph) you can
precisely see where the Go compiler is spending time, on a beefy machine
the compilation takes of the zstd C library takes ~40s. For reference
compiling the sqlite3 C library takes ~50s.
- The forgejo binary (build via `make backend`) reduced its size by
835776 bytes (0.8 MiB).