feat: show size constraint of custom avatar
Uploading a custom avatar has a size constraint for both the file size
and the image dimensions. The user was only informed about these
constraints after trying to upload an avatar and receiving an error if
the uploaded image is too big.

Change this to inform the user of the size constraint before uploading
an avatar.

Closes: https://codeberg.org/forgejo/forgejo/issues/7862
Signed-off-by: Thomas Böhler <witcher@wiredspace.de>