Prevent 404 for force-pushed commits
The pull request conversation tab may show commits that have been
dropped by a force push or other form of history rewriting.
Since those are no longer part of the branch that the pull request is
based on, linking them to the file diff tab would lead to issues.
To prevent that, those now link to the regular commit view.