chore(ci): run additional tests in integration
With git 2.30 instead of git >= 2.42

These will be run once a day, when the mirror from the development
repository happens and copies the branches to the integration
repository (see .forgejo/workflows/mirror.yml).

A failure will not be blocking and will go unoticed until Codeberg is
upgraded to v12 which can send a mail a workflow fails.